requirements.txt
asgiref
Bottleneck
certifi==2021.5.30
cx-Oracle
defusedxml
diff-match-patch
dj-database-url==0.5.0
Django
django-excel==0.0.10
django-heroku==0.3.1
django-import-export-customized==1.1.1.dev0
et-xmlfile==1.1.0
gunicorn==20.1.0
jdcal==1.4.1
lml
MarkupPy
numexpr
numpy
odfpy
openpyxl
pandas
psycopg2
pyexcel
pyexcel-io
pyexcel-webio
pyexcel-xls
pyexcel-xlsx
python-dateutil
pytz
PyYAML
six
sqlparse
tablib
tabulate
texttable
typing-extensions
whitenoise
wincertstore
xlrd
xlwt
